Closed bug ID: 821208 -- Weird peer display in dungeons
Now dungeon 'peer' map visibility is dependent on whether or not the party is able to normally walk to that map tile. Tile visibility matches behavior of c64, with the exception of some weird c64 glitches which causes peculiar invisibility in some repeating maps.

Also, the party tile will only appear once, even on repeating maps.


git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2761 694b2c84-8326-496a-a978-df44a0d360f4