Use OS-specific functions to locate xu4 resources on Mac OS X.
There's some handcrafted code for finding path to certain things in xu4. While
this is fine in general, it can be potentially fragile. Mac OS X provides
functions that take care of this in a simple way so let's just use them and put
the maintenance burden on Apple.

This fix is the same as found in patch 3197823.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2843 694b2c84-8326-496a-a978-df44a0d360f4