Ensure that we update the width and height if we don't have it.
Storing this information in the XML may be redundant, but we need to be sure
that we store this information after we load the image, because we need to use
this in other places. We aren't storing depth for the moment, but it seems most
code doesn't need it at the moment.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2889 694b2c84-8326-496a-a978-df44a0d360f4