Make iOS work in landscape orientation.
This is the first attempt at getting things to work in the landscape format.
It's usable, though could probably use some polishing. The problem is that
since I have to change the layout of the views, I have to hardcode the values.
This would be OK if the values actually made sense, but it seems that they are
put into some magic machinery.

One of the things I did to help this out was that I've created an extra label
for that mimics the text from text area of the game screen. This stores the
last 256 lines (we could probably do more, but really the original game only
had 13 lines ;). You can also copy the text out. You can keep notes that way,
though people probably would expect a real interface for that.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2893 694b2c84-8326-496a-a978-df44a0d360f4