Fix crashes when fleeing combat on iOS.
The previous code was not actually halting the event loop, which leads to very
bad problems when the controller is deleted, but has been called recursively.
Right now, I'm just sleeping the thread, but I think there must be a better
solution, though it might require re-architecting controllers (at least the
combat one). The downside is that now you get the same "delay" effect when
pressing keys that you get in the desktop versions.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2897 694b2c84-8326-496a-a978-df44a0d360f4