Fix crash when peering at gems in dungeon levels that repeat (3256763)
Apply fix suggested by nbehary99 from the tracker forum. This seems to do the correct thing. To quote the posts:

"This is caused by the code that attempts to prevent the Avatar from being
displayed multiple times in a repeating level"

And

"[this fix] keeps the avatar from showing repeatedly, and shows the tile under
the avatar in the repeating sections:"

Thanks nbehary99 for the fix!

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2909 694b2c84-8326-496a-a978-df44a0d360f4