Remove the transparency index on ladders and use real transparency
The VGA and its inherited bretheren know about real transparency, so instead of
using a color to mask it, just make the images really transparent and be done
with it.

Yes, the images take up a little bit more space now, but this is really not an
issue. It seems to work OK in the VGA tileset and there's a little problem with
the VGA-ALLPNG tileset, but that requires some more knowledge in what the SDL
code is doing as it seems to be affecting other tiles in the dungeon.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2911 694b2c84-8326-496a-a978-df44a0d360f4