Fix items not showing in dungeons in iOS.
It seems the View class has a static reference to the screen. This is fine, but
since this value changes in iOS, items may not get shown. We now update this
value on iOS. This probably could be made to not be a static reference, but I
don't want to investigate the performance issues of doing that at this time.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2912 694b2c84-8326-496a-a978-df44a0d360f4