Restored ladder PNGs, and made them transparent while retaining the index of transparency. Dungeon graphics still have issues with natural transparency and must take advantage of the transparency hack algorithm. This will hopefully be fixed soon.
The optional transparency algorithm for black-background tiles has been renamed to include the word "hack", and has also been refined and corrected. It has been extracted from the setTransparencyIndex command. All calls to this function have been cleaned up. Whether or not this gets used is based on a user parameter, as well as a graphics.xml field fixup="blackBackgroundHack". The hack also is performed pre-scaling, to reduce weird artifacts and 'invisible cloak' syndrome. The additional parameters may be fixed, once a final look has been established.

Transparency effects have been tested with all graphics modes, with success. Things look a little odd in dungeons if a hack tileset is used, and the halo size is more than 1. This will probably be fixed fairly soon as well.


git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2923 694b2c84-8326-496a-a978-df44a0d360f4