Updated Image manager to be more forgiving about missing files, and to not auto-default to EGA if the VGA upgrade is missing.
Now the graphic sets will cascade through the hierarchy of "extends" until it finds the first successful image. Thus it is possible to use Hires or VGA-PNG without having the upgrade, and it will still extract the dungeon graphics etc associated with VGA.

Eventually, I want to define a graphics mode as a container of these image sets, and find some way to efficiently let the user select which visual features are wanted. This is an incremental process...

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2924 694b2c84-8326-496a-a978-df44a0d360f4