Fix bugs where spells were inaccessible on iOS and pop location.
We had an issue when, if you had all the spells mixed, that you could not press
the buttons. It is impossible to cast *all* spells in all contexts, so there's
no point to show them all, plus it leads to losing a turn if you choose a spell
in the wrong context. We now only show the spells that make sense in that
context. We aren't checking the current player's MP, so that can lead to
issues, but keeps the experience consistent.

The direction popup now pops up in the correct position as well.

git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@2939 694b2c84-8326-496a-a978-df44a0d360f4