Stop Views from updating the screen themselves.
This change makes it so that screenRedrawScreen and screenRedrawTextArea are the only means of causing screen redraws (at least for SDL.) This means the game *should* be drawing only once per frame, rather than every time an ImageView or TextView gets updated. Effects noticed in testing include the disappearance of flicker in menus, and improved performance overall.

However, there may be side-effects in areas of the game where screenRedrawScreen don't get called and the Views are being expected to do the work. I didn't encounter any of this while testing, but that doesn't mean it doesn't happen. If it does though, it should be easy enough to fix.

This change should have no effect at all on iOS.



git-svn-id: https://svn.code.sf.net/p/xu4/code/trunk@3056 694b2c84-8326-496a-a978-df44a0d360f4